requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.6.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.1
django-cors-headers==3.4.0
djangorestframework==3.11.1
idna==2.10
isort==4.3.21
joblib==0.16.0
JPype1==1.0.2
konlpy==0.5.2
krwordrank==1.0.3
lazy-object-proxy==1.4.3
lxml==4.5.2
mccabe==0.6.1
numpy==1.19.1
oauthlib==3.1.0
pylint==2.5.3
PySocks==1.7.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sqlparse==0.3.1
threadpoolctl==2.1.0
toml==0.10.1
tweepy==3.9.0
urllib3==1.25.10
wrapt==1.12.1
